# db_api/swagger.yaml
swagger: '2.0'
info:
  description: Api calls
  version: 1.0.0
  title: DB APi routes
paths:
  /user/:
    post:
      tags:
        - user
      summary: Create a user and return the newly added user
      description: ''
      operationId: CreateUser
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
    get:
      tags:
        - user
      summary: Get all users
      description: ''
      operationId: get_users
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
                $ref: '#/definitions/User'
  '/user/{user_id}':
    get:
      tags:
        - user
      summary: Get user by id
      description: ''
      operationId: get_user_by_id
      produces:
        - application/json
      parameters:
        - name: user_id
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '404':
          description: User not found
    put:
      tags:
        - user
      summary: Update user and return the newly updated user
      operationId: updateUser
      produces:
        - application/json
      parameters:
        - name: user_id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          description: Updated user object
          required: false
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '404':
          description: User not found
    delete:
      tags:
        - user
      summary: Delete user
      operationId: deleteUser
      produces:
        - application/json
      parameters:
        - name: user_id
          in: path
          required: true
          type: integer
      responses:
        '404':
          description: User not found
        '400':
          description: This user has existing votes
...
  '/vote/user/{user_id}':
    get:
      tags:
        - vote
      summary: Get all votes for user_id
      description: ''
      operationId: get_votes_for_user
      produces:
        - application/json
      parameters:
      - name: user_id
        in: path
        required: true
        type: integer
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
                $ref: '#/definitions/Vote'
definitions:
  User:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
...
  Vote:
    type: object
    properties:
      user_id:
        type: integer
      post_id:
        type: integer
      value:
        type: integer